Refactor to avoid circular reference
