+++ starlette_zipkin/__init__.py
@@ -1,12 +1,6 @@
 from starlette_zipkin.header_formatters import B3Headers, UberHeaders
-from starlette_zipkin.middleware import (
-    ZipkinConfig,
-    ZipkinMiddleware,
-    get_ip,
-    get_root_span,
-    get_tracer,
-)
-from starlette_zipkin.trace import trace
+from starlette_zipkin.middleware import ZipkinConfig, ZipkinMiddleware, get_ip
+from starlette_zipkin.trace import get_root_span, get_tracer, trace
 
 __version__ = "0.1.4"
 __all__ = [
